finished the phone numberbug and heart function

+++ users/models.py
@@ -5,14 +5,13 @@ from django.core.validators import MinValueValidator, MaxValueValidator
 from django.core.exceptions import ValidationError
 from phonenumber_field.modelfields import PhoneNumberField
 
-import uuid, os, random, string
+import uuid, os, random, string, phonenumbers
 
 #User App model starts from here
 
 def profile_image_file_path(instance, filename):
     """Generate file path for new profile image"""
     ext = filename.split('.')[-1]
-    print(ext)
     filename = f'{uuid.uuid4()}.{ext}'
 
     return os.path.join('uploads/profilePicture/', filename)
@@ -56,7 +55,7 @@ class UserStudent(models.Model):
     """Custom userStudent model that stores Student information"""
 
     user = models.OneToOneField(UserType, on_delete=models.CASCADE)
-    phone = PhoneNumberField()
+    phone = PhoneNumberField(region='US')
     university = models.CharField(max_length=50)
     classYear = models.IntegerField(validators=[
                     MinValueValidator(2010, "Minimum year 2010"), 
@@ -76,6 +75,15 @@ class UserStudent(models.Model):
     def clean(self):
         if not self.user.is_student:
             raise ValidationError({'user': ValidationError(('User is not a student!'), code='invalid')})
+        
+        try:
+            phone = phonenumbers.parse(str(self.phone), None)
+            if phone.country_code != 1:
+                # print(phone.country_code)
+                # print(type(phone.country_code))
+                raise ValidationError({'phone': ValidationError(('Currently we accept only USA Numbers!'), code='invalid phone')})
+        except phonenumbers.NumberParseException:
+            pass
 
     def __str__(self):
         return self.user.user.username
@@ -85,7 +93,7 @@ class UserLandLord(models.Model):
     """Custom userLandLord model that stores Seller information"""
 
     user = models.OneToOneField(UserType, on_delete=models.CASCADE)
-    phone = PhoneNumberField()
+    phone = PhoneNumberField(region='US')
     emailVerified = models.BooleanField(default=False)
     phoneVerified = models.BooleanField(default=False)
     profilePicture = models.ImageField(upload_to=profile_image_file_path)
@@ -94,6 +102,15 @@ class UserLandLord(models.Model):
     def clean(self):
         if not self.user.is_landlord:
             raise ValidationError({'user': ValidationError(('User is not a Seller!'), code='invalid')})
+
+        try:
+            phone = phonenumbers.parse(str(self.phone), None)
+            if phone.country_code != 1:
+                # print(phone.country_code)
+                # print(type(phone.country_code))
+                raise ValidationError({'phone': ValidationError(('Currently we accept only USA Numbers!'), code='invalid phone')})
+        except phonenumbers.NumberParseException:
+            pass
 
     def __str__(self):
         return self.user.user.username
@@ -109,7 +126,8 @@ def auto_delete_seller_profile_pic_on_modified(sender, instance, **kwargs):
         alreadyExists = UserLandLord.objects.filter(pk=instance.pk).exists()
         if alreadyExists:
             oldFile = UserLandLord.objects.get(pk=instance.pk)
-            if str(oldFile.profilePicture) != str(instance.profilePicture):
+            # +41524204242
+            if str(oldFile.profilePicture) != str(instance.profilePicture) and (str(oldFile.profilePicture) != ''):
                 if os.path.isfile(oldFile.profilePicture.path):
                     os.remove(oldFile.profilePicture.path)
 
@@ -133,7 +151,7 @@ def auto_delete_student_profile_pic_on_modified(sender, instance, **kwargs):
         alreadyExists = UserStudent.objects.filter(pk=instance.pk).exists()
         if alreadyExists:
             oldFile = UserStudent.objects.get(pk=instance.pk)
-            if str(oldFile.profilePicture) != str(instance.profilePicture):
+            if str(oldFile.profilePicture) != str(instance.profilePicture) and (str(oldFile.profilePicture) != ''):
                 if os.path.isfile(oldFile.profilePicture.path):
                     os.remove(oldFile.profilePicture.path)
